pyproject.toml
[project]
name = "itch_jam"
version = "1.5"
description = "Jam crawler"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "click-extra>=5.0.2",
    "cloup>=3.0.7",
    "deepl>=1.22.0",
    "html2text>=2025.4.15",
    "jinja2>=3.1.6",
    "lingua-language-detector>=2.1.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
]
